API/pyproject.toml
[tool.poetry]
name = "locationreg"
version = "0.1.0"
description = ""
authors = ["Patrick Stünkel <webminz@googlemail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.8"}
minio = "^7.2.15"
psycopg2-binary = "^2.9.10"
pydantic = "^2.10.6"
dynamic-frcm = "^1.0.1"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
locationreg_server = "locationreg.main:main"

